Revert "Feat/trips page"
