+++ react-client/src/components/event-list.jsx
@@ -8,9 +8,7 @@ const eventTitleStyle = {
   marginBottom: 15,
   width: 680
 };
-const handleSubmit = () => {
 
-};
 const EventList = (props) => (
   <div id='event-list'>
   <MuiThemeProvider>
@@ -19,9 +17,6 @@ const EventList = (props) => (
       style={eventTitleStyle}
     />
   </MuiThemeProvider>
-    <form onSubmit={handleSubmit.bind(this)}>
-      <input type="file" name="image"></input><button type="submit">Upload</button>
-    </form>
     <ol style={{overflow: 'auto', width: 680, height: 180}} >
       {props.events.map((event, index) => <EventEntry
         key={index}
