+++ react-client/src/components/search_grid_tile.jsx
@@ -12,7 +12,7 @@ class SearchGridTile extends React.Component {
     super(props);
     this.state = {
       open: false,
-    };
+    }
     this.handleOpen = this.handleOpen.bind(this);
     this.handleClose = this.handleClose.bind(this);
     this.clickImage = this.clickImage.bind(this);
