+++ server/index.js
@@ -2,7 +2,7 @@ var express = require('express');
 var bodyParser = require('body-parser');
 var cookieParser = require('cookie-parser');
 var session = require('express-session');
-var db = require('../database-mongo/models.js')
+var db = require('../database-mongo/index.js');
 var bodyParser = require('body-parser');
 var request = require('request');
 var passport = require('./middleware/initGoogle.js');
