+++ server/middleware/initGoogle.js
@@ -1,12 +1,12 @@
 let GoogleStrategy = require('passport-google-oauth20').Strategy;
-// let googleCredentials = require('./authConfig.js');
+let googleCredentials = require('./authConfig.js');
 let passport = require('passport');
 let db = require('../../database-mongo/index.js');
 
 passport.use(new GoogleStrategy({
-  clientID: process.env.GOOGLE_CLIENT_ID,
-  clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-  callbackURL: process.env.GOOGLE_CALLBACK_URL
+  clientID: googleCredentials.googleClientId,
+  clientSecret: googleCredentials.googleClientSecret,
+  callbackURL: googleCredentials.googleCallbackURL
 },
 function(accesstoken, refreshToken, profile, done) {
   // console.log('this is the accesstoken', accesstoken);
